backend/requirements-api.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
waitress==3.0.0
openai==1.12.0
numpy==1.26.4
pandas==2.2.0
chromadb==0.3.21
ollama==0.6.1
pymupdf==1.24.1
fpdf2==2.7.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4
unstructured==0.10.30
openpyxl==3.1.2
